Trunk GS IL: - Fix for siege skills when switching subclass - Little DimensionalRiftManager fix
git-svn-id: http://svn.l2jfree.com/l2jfree@2094 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_IL/src/main/java/net/sf/l2j/gameserver/instancemanager/DimensionalRiftManager.java
@@ -275,20 +275,20 @@ public class DimensionalRiftManager
     public boolean checkIfInRiftZone(int x, int y, int z, boolean ignorePeaceZone)
     {
         if (ignorePeaceZone)
-            return getRooms().get(RoomType.DimensionalRift).get(1).checkIfInZone(x, y, z);
+            return getRooms().get(RoomType.DimensionalRift).get((byte)1).checkIfInZone(x, y, z);
         else
-            return getRooms().get(RoomType.DimensionalRift).get(1).checkIfInZone(x, y, z)
-                    && !getRooms().get(RoomType.Start).get(1).checkIfInZone(x, y, z);
+            return getRooms().get(RoomType.DimensionalRift).get((byte)1).checkIfInZone(x, y, z)
+                    && !getRooms().get(RoomType.Start).get((byte)1).checkIfInZone(x, y, z);
     }
 
     public boolean checkIfInPeaceZone(int x, int y, int z)
     {
-        return getRooms(RoomType.Start).get(1).checkIfInZone(x, y, z);
+        return getRooms(RoomType.Start).get((byte)1).checkIfInZone(x, y, z);
     }
 
     public void teleportToWaitingRoom(L2PcInstance player)
     {
-        Location loc = getRooms(RoomType.Start).get(1).getTeleport();
+        Location loc = getRooms(RoomType.Start).get((byte)1).getTeleport();
         player.teleToLocation(loc, false);
     }
